requirements.txt
tkinter
openai
python-dotenv
tkcalendar
'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'